API/requirements.txt
ffmpeg-python==0.2.0
Flask==1.1.2
google-cloud-speech==2.0.1
google-cloud-storage==1.35.0
Werkzeug==1.0.1
gunicorn==20.0.4
flask-cors==3.0.10